cheryl: Update adreno stack
As the adreno userspace binaries were built for Android 8.0,
when -ffixed-x18 was not guarteed to be used, which had been
ensured for Android versions newer than P for the shadow
call stack (SCS) feature.
For such old version vendor blob, there was a workaround
in com_android_internal_os_ZygoteInit.cpp which has been removed
now.
    https://android-review.googlesource.com/c/platform/frameworks/base/+/2518435

After that removal, there is a crash on libbluetooth_jni.so and
libnfc_nci_jni.so

Take the blobs from OnePlus 5/5T because it looks like the most "recent"
stack among msm8998 devices and even when compared with other cool
sdm660 devices running 4.4 kernel.

Reference:
    https://android-review.googlesource.com/c/device/linaro/hikey/+/2902255
Change-Id: I4392fdef0c2062e18b9a09fcd2bfd4759b57b93c